pyproject.toml
[tool.poetry]
name = "scrap"
version = "0.1.0"
description = ""
authors = ["Prasanth Thiagarajan <prasanth33460@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.27.1"
google-api-python-client = "^2.154.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pynput = "^1.7.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
